Version 1.3.0 of the AWS Java SDK
This release adds support for working with the new Amazon DynamoDB
service, including a high-level API for annotating Java classes with
instructions on mapping them to Amazon DynamoDB tables, API updates
for the Amazon SNS client, and general SDK improvements such as
.equals and .hashCode implementations for model objects.

Full release notes:

http://aws.amazon.com/releasenotes/Java/5581335213116164

+++ src/main/java/com/amazonaws/services/autoscaling/model/DeleteAutoScalingGroupRequest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2011 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright 2010-2012 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * 
  * Licensed under the Apache License, Version 2.0 (the "License").
  * You may not use this file except in compliance with the License.
@@ -164,10 +164,35 @@ public class DeleteAutoScalingGroupRequest extends AmazonWebServiceRequest {
     public String toString() {
         StringBuilder sb = new StringBuilder();
         sb.append("{");
-        sb.append("AutoScalingGroupName: " + autoScalingGroupName + ", ");
-        sb.append("ForceDelete: " + forceDelete + ", ");
+        if (autoScalingGroupName != null) sb.append("AutoScalingGroupName: " + autoScalingGroupName + ", ");
+        if (forceDelete != null) sb.append("ForceDelete: " + forceDelete + ", ");
         sb.append("}");
         return sb.toString();
+    }
+    
+    @Override
+    public int hashCode() {
+        final int prime = 31;
+        int hashCode = 1;
+        
+        hashCode = prime * hashCode + ((getAutoScalingGroupName() == null) ? 0 : getAutoScalingGroupName().hashCode()); 
+        hashCode = prime * hashCode + ((isForceDelete() == null) ? 0 : isForceDelete().hashCode()); 
+        return hashCode;
+    }
+    
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj) return true;
+        if (obj == null) return false;
+    
+        if (obj instanceof DeleteAutoScalingGroupRequest == false) return false;
+        DeleteAutoScalingGroupRequest other = (DeleteAutoScalingGroupRequest)obj;
+        
+        if (other.getAutoScalingGroupName() == null ^ this.getAutoScalingGroupName() == null) return false;
+        if (other.getAutoScalingGroupName() != null && other.getAutoScalingGroupName().equals(this.getAutoScalingGroupName()) == false) return false; 
+        if (other.isForceDelete() == null ^ this.isForceDelete() == null) return false;
+        if (other.isForceDelete() != null && other.isForceDelete().equals(this.isForceDelete()) == false) return false; 
+        return true;
     }
     
 }
